Login additions
It is now setup so a user can only unsuccessfully login 3 times without penalty. On the fourth time, a message appears that the max attempts has been reached and the user can no longer attempt to login. If the user fails 1-3 times, a message appears that says either the username or password is invalid